Changes in how the application is setup
Better controlling for routes, added repositories, and removed a few model classes.
