+++ Server.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const cors = require('cors')
 const app = express()
-const formRoutes = require('./src/server/controllers/FormController')
+const formRoutes = require('./src/server/routes/formRoutes')
 
 app.use(cors({
     origin: ['http://localhost:5173', 'http://localhost:5000'],
