+++ src/server/models/Shteti.js
@@ -1,39 +1,36 @@
-const BaseModel = require('./BaseModel');
+//Will be removed in the future, placeholder for how classes should look like
+//This class has only static data, it cant be changed cause only a certain number of countries exist
+//Only add class models that have create, update, delete operations
+//This class is just for reference
+//This class is not used in the code
+class Shteti {
+    #shtetiID;
+    #emriShtetit;
+    #iso_kodi;
 
-/**
- * ShtetiModel - Handles country-related database operations
- * Example model extending BaseModel
- */
-class Shteti extends BaseModel {
-    constructor() {
-        super('shteti', 'shtetiID');
+    constructor(shtetiID, emriShtetit, iso_kodi) {
+        this.#shtetiID = shtetiID;
+        this.#emriShtetit = emriShtetit;
+        this.#iso_kodi = iso_kodi;
     }
 
-    /**
-     * Find country by ISO code
-     * @param {string} isoCode - ISO country code
-     * @returns {Promise<Object|null>} - Country object or null
-     */
-    async findByIsoCode(isoCode) {
-        return await this.findOneByField('iso_kodi', isoCode);
+    getShtetiID() {
+        return this.#shtetiID;
     }
 
-    /**
-     * Search countries by name (case insensitive)
-     * @param {string} name - Country name to search
-     * @returns {Promise<Array>} - Array of matching countries
-     */
-    async searchByName(name) {
-        const query = `SELECT * FROM ${this.tableName} WHERE LOWER(emri_shtetit) LIKE LOWER(?) ORDER BY emri_shtetit`;
-        return await this.query(query, [`%${name}%`]);
+    getEmriShtetit() {
+        return this.#emriShtetit;
+    }   
+    getIso_kodi() {
+        return this.#iso_kodi;
     }
 
-    /**
-     * Get all countries ordered by name
-     * @returns {Promise<Array>} - Array of all countries
-     */
-    async getAllOrderedByName() {
-        return await this.findAll({ orderBy: 'emri_shtetit' });
+    setShtetiID(shtetiID) {
+        this.#shtetiID = shtetiID;
+    }
+
+    setIsokodi(iso_kodi) {
+        this.#iso_kodi = iso_kodi;
     }
 }
