+++ src/server/models/index.js
@@ -1,8 +1,6 @@
 // Export all models for easy importing
-const BaseModel = require('./BaseModel');
 const Shteti = require('./Shteti');
 
 module.exports = {
-    BaseModel,
     Shteti
 };
