+++ src/server/repositories/index.js
@@ -1,0 +1,7 @@
+const BaseRepository = require('./BaseRepository');
+const ShtetiRepository = require('./ShtetiRepository');
+
+module.exports = {
+    BaseRepository,
+    ShtetiRepository
+};
